Add color coding for dimmed todo entries
Changed todo entry color coding to allow any number of ’!’s so that
multiple ’!’s can be used to denote more important tasks. Possibly add
more bright colors for those lately. Bright, primary red is a good
incentive to finish the task to save my eyes, yes?

Also changed default colors and fonts to sensible values, so they can be
reverted to with a single $font or $color. This required getting rid of
the border, as otherwise it uses the default color, but I never liked
that anyway.