Fix my-setq-append and rename to my-append
Now actually works as well as doesn’t require LIST-VAR to already be
defined; i.e. it can be used in the :init section of packages, although
I no longer need this as much as I did before due to moving most
configuration settings to load after the package with :config.